requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
pytest==8.2.0
pytest-xdist==3.5.0
pytest-html==4.1.1
allure-pytest==2.13.5
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.1
Faker==24.11.0
pytest-rerunfailures
